removed mysql configuration in settings.py and test.py file
