+++ application/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 # class FriendHubUser(models.Model):
-#     username = models.CharField(max_length=30)
+#     user = models.CharField(max_length=30)
 #     email = models.CharField(max_length=30)
 
 #     def __str__(self) -> str:
