+++ application/tests.py
@@ -1,3 +1,8 @@
 from django.test import TestCase
+from .models import FriendHubUser  # Import your model here
 
-# Create your tests here.
+class CreateUserTestCase(TestCase):
+    def test_user_creation(self):
+        instance = FriendHubUser.objects.create(username='example_username', email='example_email')
+        self.assertIsInstance(instance, FriendHubUser)
+        self.assertEqual(instance.username, 'example_username')
